Node.js fixes and documenting limitations
This PR is the first in a series of fixes for bringing this application
up to par with the latest Node.js versions. Specifically this PR
adds a .nvmrc file that suggests which version of Node.js to use, and
documents the suggested version in the getting started guide.

Additionally, the existing package-lock.json was created with an
old version of npm, this PR checks in a new version of the file that
was created with a newer version of npm.

Signed-off-by: Steve Martinelli <steve.martinelli@shopify.com>