clean up tutorial docs
This PR consolidates the documentation used in the tutorial
section of this repo. The main tutorial with the steps to
install the embedded application are the most updated.

Some clean up was done to the extension related docs, however,
future PRs will clean them up further.

This PR also adds some much needed images and wording around
what the reader should be looking for at the end of the tutorial.

Closes: #34
Closes: #35

Signed-off-by: Steve Martinelli <steve.martinelli@shopify.com>